git-svn-id: https://svn.apache.org/repos/asf/sis/trunk@1551743 13f79535-47bb-0310-9956-ffa450edef68

+++ storage/sis-netcdf/src/main/java/org/apache/sis/internal/netcdf/impl/ChannelDecoder.java
@@ -252,7 +252,7 @@ public final class ChannelDecoder extends Decoder {
      *         one of the expected constants.
      */
     private static String tagName(final int tag) {
-        final int key;
+        final short key;
         switch (tag) {
             case DIMENSION: key = Vocabulary.Keys.Dimensions; break;
             case VARIABLE:  key = Vocabulary.Keys.Variables;  break;
